Set scheduler_iteration=10 Torque server attribute
This fix wants to address a problem with the Torque scheduler in the
scenario when the cluster has no compute nodes available.
Jobs get scheduled and go in Q state, but even when new compute nodes
become available to the cluster, the jobs remain in Q state.
This has a major impact to customers that see the cluster scaling up to
the max size but without any job running.

From tested approaches and according to Adaptive, one way to recover
from this situation is to set the scheduler iteration to a more
aggressive time interval, e.g. 10 seconds. Default is 600 secs, and it's
reset at every pbs_server restart, so each time a new node was added to
the cluster.